+++ src/app/leader-auditionee/leader-auditionee.component.html
@@ -11,12 +11,6 @@
 <h3>Auditionee</h3>
 <input (keyup)="onKeyAuditionee($event)" type="text" id="auditionee" name="auditionee" placeholder="Name an auditionee"  required/>
 
-
-
-<app-judgement></app-judgement>
-<input type="button" id="more_fields" (click)="add_criteria();" value="Add another comment"/>
-
-
 <div>
 	<ng-template #target></ng-template>
 </div>
